games/requirements.txt
Django==3.1.6
psycopg2==2.8.6
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
django-phonenumber-field[phonenumbers]==5.1.0
djangorestframework-simplejwt==4.1.4
Pillow==8.2.0
django_filter==2.4.0
Markdown==3.3.4
django-url-filter==0.3.15